+++ features/profile/components/organisms/other-page.tsx
@@ -5,7 +5,6 @@ import { useState } from 'react';
 import { ProfileProps } from '@/app/profile/[id]/page';
 import BadgeIcon from '@/components/atoms/icons/badge-icon';
 import StatisticsIcon from '@/components/atoms/icons/statistics-icon';
-import { UserImageIcon } from '@/components/atoms/icons/user-image-icon';
 import { Tab, TabItem } from '@/components/molecules';
 import { Calendar } from '@/features/main';
 import { css } from '@/styled-system/css';
@@ -13,38 +12,19 @@ import { flex } from '@/styled-system/patterns';
 import { ProfileType } from '@/types/profileType';
 
 import FollowButton from '../atoms/follow-button';
+import ProfileContainer from './profile-container';
 
 export function OtherPage({
   profileData,
 }: {
   profileData: ProfileProps['data'];
 }) {
-  const [selectedTab, setSelectedTab] = useState<ProfileType>('statistics');
+  const [selectedTab, setSelectedTab] = useState<ProfileType>('record');
 
   return (
     <div>
       <section className={profileContainer}>
-        <div className={inforWrapper}>
-          <UserImageIcon />
-          <div>
-            <div className={nameStyles}>{profileData.nickname}</div>
-            <div className={introStyles}>{profileData.introduction}</div>
-          </div>
-        </div>
-        <div className={friendStyles.container}>
-          <div className={friendStyles.item}>
-            <div className={friendStyles.type}>팔로워</div>
-            <div className={friendStyles.count}>
-              {profileData.followerCount}
-            </div>
-          </div>
-          <div className={friendStyles.item}>
-            <div className={friendStyles.type}>팔로잉</div>
-            <div className={friendStyles.count}>
-              {profileData.followingCount}
-            </div>
-          </div>
-        </div>
+        <ProfileContainer profileData={profileData} />
         <div className={buttonContainer}>
           <FollowButton />
         </div>
@@ -105,48 +85,6 @@ const profileContainer = flex({
   gap: '20px',
   padding: '20px',
 });
-
-const inforWrapper = flex({
-  alignItems: 'flex-start',
-  gap: '21px',
-  alignSelf: 'stretch',
-});
-
-const nameStyles = css({
-  textStyle: 'heading3',
-  fontWeight: '600',
-  color: 'text.normal',
-});
-
-const introStyles = css({
-  textStyle: 'body2.normal',
-  fontWeight: '500',
-  color: 'text.alternative',
-  width: '254px',
-});
-
-const friendStyles = {
-  container: flex({
-    alignItems: 'center',
-    gap: '40px',
-    alignSelf: 'stretch',
-  }),
-  item: flex({
-    direction: 'column',
-    alignItems: 'flex-start',
-    gap: '2px',
-  }),
-  type: css({
-    textStyle: 'label2',
-    fontWeight: '500',
-    color: 'text.alternative',
-  }),
-  count: css({
-    textStyle: 'body2.normal',
-    fontWeight: '600',
-    color: 'text.normal',
-  }),
-};
 
 const buttonContainer = flex({
   width: '100%',
